+++ documentation/plugins/blog-plugin.js
@@ -210,6 +210,17 @@ async function blogPluginExtended(...pluginArgs) {
                         JSON.stringify(metadata, null, 2),
                     );
 
+                    const tagsProp = Object.values(blogTags).map((tag) => ({
+                        label: tag.label,
+                        permalink: tag.permalink,
+                        count: tag.items.length,
+                    }));
+
+                    const tagsPropPath = await createData(
+                        `${utils.docuHash(`${blogTagsListPath}-tags`)}.json`,
+                        JSON.stringify(tagsProp, null, 2),
+                    );
+
                     addRoute({
                         path: permalink,
                         component: "@theme/BlogListPage",
@@ -226,6 +237,7 @@ async function blogPluginExtended(...pluginArgs) {
                                     : items,
                             ),
                             metadata: aliasedSource(pageMetadataPath),
+                            tags: aliasedSource(tagsPropPath),
                         },
                     });
                 }),
@@ -311,6 +323,19 @@ async function blogPluginExtended(...pluginArgs) {
                             JSON.stringify(metadata, null, 2),
                         );
 
+                        const tagsProp = Object.values(blogTags).map((tag) => ({
+                            label: tag.label,
+                            permalink: tag.permalink,
+                            count: tag.items.length,
+                        }));
+
+                        const tagsPropPath = await createData(
+                            `${utils.docuHash(
+                                `${blogTagsListPath}-tags`,
+                            )}.json`,
+                            JSON.stringify(tagsProp, null, 2),
+                        );
+
                         addRoute({
                             path: metadata.permalink,
                             component: "@theme/BlogTagsPostsPage",
@@ -318,6 +343,7 @@ async function blogPluginExtended(...pluginArgs) {
                             modules: {
                                 items: blogPostItemsModule(items),
                                 tag: aliasedSource(tagPropPath),
+                                tags: aliasedSource(tagsPropPath),
                                 listMetadata: aliasedSource(listMetadataPath),
                             },
                         });