+++ examples/multi-tenancy-strapi/src/components/product/item.tsx
@@ -1,5 +1,4 @@
 import { DeleteButton } from "@refinedev/antd";
-
 import { EditOutlined } from "@ant-design/icons";
 import { Card } from "antd";
 
@@ -14,7 +13,7 @@ type ProductItemProps = {
 const { Meta } = Card;
 
 export const ProductItem: React.FC<ProductItemProps> = ({ item, editShow }) => {
-    const image = item.image ? API_URL + item.image.url : "./error.png";
+    const image = item.image ? API_URL + item.image.url : "/error.png";
 
     return (
         <Card