+++ packages/antd/src/contexts/themedLayoutContext/index.tsx
@@ -4,14 +4,17 @@ import { IThemedLayoutContext } from "./IThemedLayoutContext";
 
 export const ThemedLayoutContext = React.createContext<IThemedLayoutContext>({
     siderVisible: false,
-    drawerSiderVisible: true,
+    drawerSiderVisible: false,
 });
 
-export const ThemedLayoutContextProvider: React.FC<{ children: ReactNode }> = ({
-    children,
-}) => {
+export const ThemedLayoutContextProvider: React.FC<{
+    children: ReactNode;
+    initialSiderCollapsed?: boolean;
+}> = ({ children, initialSiderCollapsed }) => {
     const [siderVisible, setSiderVisible] = useState(false);
-    const [drawerSiderVisible, setDrawerSiderVisible] = useState(true);
+    const [drawerSiderVisible, setDrawerSiderVisible] = useState(
+        initialSiderCollapsed ?? false,
+    );
 
     return (
         <ThemedLayoutContext.Provider